pug_or_ugh_v2/requirements.txt
astroid==2.2.5
certifi==2019.6.16
colorama==0.4.1
Django==2.2.4
django-multiselectfield==0.1.8
djangorestframework==3.10.2
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wincertstore==0.2
wrapt==1.11.2
